+++ libpgmodeler_ui/src/baseform.cpp
@@ -51,7 +51,8 @@ void BaseForm::resizeForm(QWidget *widget)
 	max_h = screen->size().height() * 0.70;
 	dpi_factor = screen->logicalDotsPerInch() / 96.0f;
 
-	if(dpi_factor < 1.0f) dpi_factor = 1.0f;
+	if(dpi_factor <= 1.01f)
+		dpi_factor = 1.0f;
 
 	vbox->setContentsMargins(2,2,2,2);
 
@@ -89,7 +90,7 @@ void BaseForm::resizeForm(QWidget *widget)
 
 	// If the current height is greater than the widget's minimum height we will use a medium value
 	if(curr_h > min_size.height() && min_size.height() < max_h)
-		curr_h = (curr_h + min_size.height())/2;
+		curr_h = (curr_h + min_size.height())/2.5;
 	//Using the maximum height if the widget's minimum height exceeds the maximum allowed
 	else if(min_size.height() >= max_h)
 		curr_h = max_h;
@@ -101,7 +102,16 @@ void BaseForm::resizeForm(QWidget *widget)
 							((buttons_lt->contentsMargins().top() +
 								buttons_lt->contentsMargins().bottom()) * 6);
 
-	this->setMinimumSize(curr_w * dpi_factor, curr_h * dpi_factor);
+	curr_w *= dpi_factor;
+	curr_h *= dpi_factor;
+
+	if(curr_w > screen->size().width())
+		curr_w = screen->size().width() * 0.80;
+
+	if(curr_h > screen->size().height())
+		curr_h = screen->size().height() * 0.80;
+
+	this->setMinimumSize(curr_w, curr_h);
 	this->resize(this->minimumSize());
 }