+++ libpgmodeler_ui/src/generalconfigwidget.h
@@ -50,7 +50,8 @@ class GeneralConfigWidget: public BaseConfigWidget, public Ui::GeneralConfigWidg
 		HintTextWidget *simp_obj_creation_ht, *confirm_validation_ht, *corner_move_ht,
 		*save_last_pos_ht, *invert_rangesel_ht, *disable_smooth_ht,
 		*hide_ext_attribs_ht, *hide_table_tags_ht, *hide_rel_name_ht,
-		*code_completion_ht, *use_placeholders_ht, *min_obj_opacity_ht;
+		*code_completion_ht, *use_placeholders_ht, *min_obj_opacity_ht,
+		*autosave_ht, *op_history_ht, *ui_language_ht, *grid_size_ht, *use_curved_lines_ht;
 
 		ColorPickerWidget *line_numbers_cp, *line_numbers_bg_cp, *line_highlight_cp;