+++ libpgmodeler_ui/src/pgmodeleruins.cpp
@@ -283,7 +283,7 @@ namespace PgModelerUiNS {
 		dpi_factor = screen->logicalDotsPerInch() / 96.0f;
 
 		//If the dpi_factor is unchanged (1) we keep the dialog original dimension
-		if(dpi_factor <= 1)
+		if(dpi_factor <= 1.01f)
 			return;
 
 		max_h = screen->size().height() * 0.70;
@@ -302,12 +302,21 @@ namespace PgModelerUiNS {
 
 		// If the current height is greater than the widget's minimum height we will use a medium value
 		if(curr_h > min_size.height() && min_size.height() < max_h)
-			curr_h = (curr_h + min_size.height())/2;
+			curr_h = (curr_h + min_size.height())/2.5;
 		//Using the maximum height if the widget's minimum height exceeds the maximum allowed
 		else if(min_size.height() >= max_h)
 			curr_h = max_h;
 
-		widget->setMinimumSize(curr_w * dpi_factor, curr_h * dpi_factor);
+		curr_w *= dpi_factor;
+		curr_h *= dpi_factor;
+
+		if(curr_w > screen->size().width())
+			curr_w = screen->size().width() * 0.80;
+
+		if(curr_h > screen->size().height())
+			curr_h = screen->size().height() * 0.80;
+
+		widget->setMinimumSize(curr_w, curr_h);
 		widget->resize(widget->minimumSize());
 	}
 }